start-mcp-agentrun-knowledgebase/src/code/requirements.txt
mcp==1.25.0
starlette==0.46.1
agentrun-sdk>=0.0.25